+++ lib/util/web_view_mixin.dart
@@ -1,7 +1,43 @@
-mixin MyMixin on State<MyWidget> {
-  @override
-  void initState() {
-    // Initialization code goes here
-    super.initState();
+import 'package:flutter/material.dart';
+import 'package:humhub/util/manifest.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import 'extensions.dart';
+
+mixin WebViewMixin {
+  WebViewController getWebViewController(Manifest manifest) {
+    WebViewController controller = WebViewController()
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..setBackgroundColor(HexColor(manifest.backgroundColor))
+      ..setNavigationDelegate(
+        NavigationDelegate(
+          onProgress: (int progress) {
+            debugPrint('WebView is loading (progress : $progress%)');
+          },
+          onPageStarted: (String url) {
+            debugPrint('Page started loading: $url');
+          },
+          onPageFinished: (String url) async {
+            /*final String cookies = await webViewController
+                .runJavaScriptReturningResult('document.cookie') as String;*/
+            debugPrint('Page finished loading: $url');
+          },
+          onWebResourceError: (WebResourceError error) {},
+          onNavigationRequest: (NavigationRequest request) {
+            if (!request.url.startsWith(manifest.baseUrl)) {
+              launchUrl(Uri.parse(request.url),
+                  mode: LaunchMode.externalApplication);
+              return NavigationDecision.prevent;
+            }
+            return NavigationDecision.navigate;
+          },
+        ),
+      )
+      ..loadRequest(
+        Uri.parse(manifest.startUrl),
+      );
+
+    return controller;
   }
 }
